test(e2e): compose provider journeys from typed facts
