+++ tests/e2e/journey_cases.py
@@ -2,12 +2,12 @@
 
 import json
 import os
-import tomllib
 from collections.abc import Callable, Iterable
 from pathlib import Path
 from typing import TypeVar
 from urllib.parse import urlparse
 
+import tomllib
 from journey_types import (
     CargoEvidence,
     JourneyCase,
@@ -17,8 +17,10 @@ from journey_types import (
     ObservableAssertion,
     ProductJourneyCase,
     ProviderJourneyCase,
+    ProviderJourneyReplayFacts,
     ProviderWorld,
     PytestEvidence,
+    SlackChannelFixture,
 )
 from provider_capability_inventory import EMULATE_SUPPORTED_TOOLS
 
@@ -110,9 +112,23 @@ def unreset_mutating_tools() -> frozenset[str]:
             if tool_name not in _MUTATING_PROVIDER_TOOLS:
                 unreset.add(tool_name)
     return frozenset(unreset)
+
+
 _REPEAT_AFTER_RESET = {
     "qa_5d_slack_strategy_doc_answer",
     "qa_10f_slack_mention_encoding",
+}
+_PROVIDER_REPLAY_FACTS = {
+    "qa_7c_slack_bug_logger_routine": ProviderJourneyReplayFacts(
+        google_spreadsheet_id="sheet_reborn_bug_tracker"
+    ),
+    "qa_7e_slack_bug_sheet_delivery": ProviderJourneyReplayFacts(
+        google_spreadsheet_id="sheet_reborn_bug_tracker"
+    ),
+    "qa_10e_slack_error_honesty": ProviderJourneyReplayFacts(
+        slack_channel=SlackChannelFixture.MISSING,
+        expected_capability_failure="channel_not_found",
+    ),
 }
 
 _PYTEST_PROVIDER_EVIDENCE = PytestEvidence(
@@ -181,6 +197,9 @@ def _provider_journey_cases() -> tuple[ProviderJourneyCase, ...]:
                     ObservableAssertion.PROVIDER_READBACK,
                 ),
                 evidence=_PYTEST_PROVIDER_EVIDENCE,
+                replay=_PROVIDER_REPLAY_FACTS.get(
+                    case_id, ProviderJourneyReplayFacts()
+                ),
                 repeat_after_reset=case_id in _REPEAT_AFTER_RESET,
             )
         )
