+++ tests/e2e/scenarios/test_reborn_qa_trace_full_path.py
@@ -8,23 +8,18 @@ the recorded model's final wording.
 
 import asyncio
 import json
-import uuid
 from collections import Counter
-from datetime import UTC, datetime, timedelta
 from pathlib import Path
-from urllib.parse import parse_qs, urlparse
+from urllib.parse import urlparse
 
 import httpx
 import pytest
-
 from emulate_provider import (
-    github_headers,
     github_json,
     google_headers,
     slack_headers,
     slack_post,
 )
-from helpers import EMULATE_GITHUB_BEARER, EMULATE_SLACK_BEARER
 from journey_cases import (
     PROVIDER_JOURNEY_RUN_IDS,
     PROVIDER_JOURNEY_RUNS,
@@ -37,16 +32,31 @@ from provider_capability_inventory import (
 )
 from provider_fault_cases import PROVIDER_FAULT_CASES
 from provider_fault_proxy import PROVIDER_FAULT_PROFILES
+from provider_journey_github import emulate_github_supports_release_writes
+from provider_journey_google import (
+    assert_google_provider_baseline,
+    assert_google_provider_outcome,
+)
+from provider_journey_slack import (
+    assert_slack_provider_baseline,
+    assert_slack_provider_outcome,
+    cleanup_slack_provider_mutations,
+)
+from provider_journey_trace import (
+    compile_provider_journey_trace,
+    load_recorded_trace,
+    recorded_provider_calls,
+    recorded_trace_uses_tool_prefix,
+)
+from provider_journey_world import build_provider_journey_world
 from provider_operation_cases import PROVIDER_OPERATION_CASES
 from provider_operation_types import ProviderOperationCase
 from reborn_webui_harness import (
     YOLO_PROFILE,
     capability_preview_payload,
-    client_action_id,
     close_reborn_server,
     create_thread,
     enable_reborn_global_auto_approve,
-    fetch_extension_oauth_requirement,
     reborn_bearer_headers,
     send_message,
     start_reborn_webui_v2_server,
@@ -56,67 +66,8 @@ from reborn_webui_harness import (
 pytest_plugins = ["reborn_webui_harness"]
 
 ROOT = Path(__file__).resolve().parents[3]
-TRACE_DIR = ROOT / "tests/fixtures/llm_traces/reborn_qa/live_canary"
-MANIFEST_PATH = TRACE_DIR / "case-manifest.json"
 
-GITHUB_RELEASE_WRITE_UNAVAILABLE = {
-    403,
-    404,
-}
-GITHUB_RELEASE_WRITE_PROBE_PAYLOAD = {"tag_name": ""}
-GOOGLE_EXTENSIONS = (
-    "gmail",
-    "google-calendar",
-    "google-drive",
-    "google-docs",
-    "google-sheets",
-    "google-slides",
-)
-GOOGLE_EXTENSION_SCOPES = {
-    "gmail": (
-        "https://www.googleapis.com/auth/gmail.readonly",
-        "https://www.googleapis.com/auth/gmail.send",
-        "https://www.googleapis.com/auth/gmail.modify",
-    ),
-    "google-calendar": (
-        "https://www.googleapis.com/auth/calendar.readonly",
-        "https://www.googleapis.com/auth/calendar.events",
-    ),
-    "google-drive": (
-        "https://www.googleapis.com/auth/drive.readonly",
-        "https://www.googleapis.com/auth/drive",
-    ),
-    "google-docs": (
-        "https://www.googleapis.com/auth/documents",
-        "https://www.googleapis.com/auth/documents.readonly",
-    ),
-    "google-sheets": (
-        "https://www.googleapis.com/auth/spreadsheets",
-        "https://www.googleapis.com/auth/spreadsheets.readonly",
-    ),
-    "google-slides": (
-        "https://www.googleapis.com/auth/presentations",
-        "https://www.googleapis.com/auth/presentations.readonly",
-    ),
-}
-GOOGLE_CUMULATIVE_SCOPES = tuple(
-    dict.fromkeys(
-        scope
-        for extension_scopes in GOOGLE_EXTENSION_SCOPES.values()
-        for scope in extension_scopes
-    )
-)
-GOOGLE_TOOL_PREFIXES = (
-    "gmail__",
-    "google-calendar__",
-    "google-docs__",
-    "google-drive__",
-    "google-sheets__",
-    "google-slides__",
-)
 PROVIDER_TOOL_NAMES = EMULATE_SUPPORTED_TOOLS
-ALL_EXTENSIONS = (*GOOGLE_EXTENSIONS, "github", "slack")
-TRACE_BOOTSTRAP_TOOLS = {"builtin__extension_search"}
 
 
 @pytest.fixture(scope="module")
@@ -189,14 +140,9 @@ async def reborn_qa_emulate_runtime(
         },
     )
     await enable_reborn_global_auto_approve(base_url)
-    slack_state = await _seed_slack_workspace(emulate_slack_server["url"])
-    await _configure_slack(base_url, slack_state)
-    await _install_extensions(base_url, ALL_EXTENSIONS)
-    for extension_id in GOOGLE_EXTENSION_SCOPES:
-        await _seed_google_account(base_url, extension_id)
-    await _seed_github_account(base_url)
-    await _seed_slack_account(base_url, emulate_slack_server["url"], slack_state)
-    await _assert_extensions_active(base_url, ALL_EXTENSIONS)
+    slack_state = await build_provider_journey_world(
+        base_url, emulate_slack_server["url"]
+    )
     try:
         yield {
             "base_url": base_url,
@@ -223,10 +169,14 @@ async def reborn_qa_emulate_provider_server(
         yield reborn_qa_emulate_runtime
     finally:
         if "slack" in services:
-            await _cleanup_slack_provider_mutations(
+            compiled = _compile_journey_case(
+                journey_case,
+                reborn_qa_emulate_runtime["slack_state"],
+            )
+            await cleanup_slack_provider_mutations(
                 reborn_qa_emulate_runtime["emulate_slack_url"],
                 reborn_qa_emulate_runtime["slack_state"],
-                journey_case.case_id,
+                recorded_provider_calls(compiled.trace, PROVIDER_TOOL_NAMES),
             )
         if reset_services:
             await resettable_emulate_provider_world.reset(reset_services)
@@ -242,9 +192,7 @@ async def reborn_provider_operation_server(
     try:
         yield reborn_qa_emulate_runtime
     finally:
-        await resettable_emulate_provider_world.reset(
-            {operation_case.provider_service}
-        )
+        await resettable_emulate_provider_world.reset({operation_case.provider_service})
 
 
 @pytest.fixture
@@ -265,621 +213,15 @@ async def reborn_provider_fault_server(
         )
 
 
-async def _seed_google_account(
-    base_url: str,
-    extension_id: str,
-) -> None:
-    expires_at = (datetime.now(UTC) + timedelta(minutes=5)).isoformat()
-    async with httpx.AsyncClient(headers=reborn_bearer_headers()) as client:
-        requirement = await fetch_extension_oauth_requirement(
-            client,
-            base_url,
-            extension_id,
-        )
-        started = await client.post(
-            f"{base_url}/api/webchat/v2/extensions/{extension_id}/setup/oauth/start",
-            json={
-                "requirement": requirement["name"],
-                "expires_at": expires_at,
-                "invocation_id": requirement["setup"].get("invocation_id"),
-            },
-            timeout=15,
-        )
-        assert started.is_success, started.text
-        started_body = started.json()
-        authorization_url = started_body["authorization_url"]
-        state = parse_qs(urlparse(authorization_url).query)["state"][0]
-
-        callback = await client.get(
-            f"{base_url}/api/reborn/product-auth/oauth/google/callback",
-            params={
-                "state": state,
-                "code": f"mock_auth_code_{extension_id.replace('-', '_')}",
-                "scope": " ".join(GOOGLE_CUMULATIVE_SCOPES),
-            },
-            headers={"Accept": "application/json"},
-            timeout=30,
-        )
-        assert callback.is_success, callback.text
-        flow_status = await client.get(
-            f"{base_url}/api/reborn/product-auth/oauth/flow/"
-            f"{started_body['flow_id']}/status",
-            params={
-                "invocation_id": started_body["callback_scope"]["invocation_id"]
-            },
-            timeout=30,
-        )
-        flow_status.raise_for_status()
-        assert flow_status.json()["status"] == "completed", flow_status.text
-        invocation_id = started_body["callback_scope"]["invocation_id"]
-        listed = await client.post(
-            f"{base_url}/api/reborn/product-auth/accounts/list",
-            json={
-                "provider": "google",
-                "requester_extension": extension_id,
-                "invocation_id": invocation_id,
-            },
-            timeout=30,
-        )
-        listed.raise_for_status()
-        accounts = listed.json()["accounts"]
-        # Re-authenticating the same Emulate identity upgrades the existing
-        # user-reusable account in its original invocation scope. Only the
-        # first flow therefore exposes a newly selectable account here.
-        if not accounts:
-            return
-        assert len(accounts) == 1, listed.text
-        selected = await client.post(
-            f"{base_url}/api/reborn/product-auth/accounts/select",
-            json={
-                "provider": "google",
-                "requester_extension": extension_id,
-                "account_id": accounts[0]["id"],
-                "invocation_id": invocation_id,
-            },
-            timeout=30,
-        )
-        selected.raise_for_status()
-
-
-async def _seed_slack_workspace(emulate_url: str) -> dict[str, str]:
-    """Create deterministic Slack data and return its provider-issued IDs."""
-    async with httpx.AsyncClient(timeout=15) as client:
-        identity = await slack_post(client, emulate_url, "auth.test")
-        users = await slack_post(client, emulate_url, "users.list")
-        by_name = {member["name"]: member for member in users["members"]}
-        channels = await slack_post(
-            client,
-            emulate_url,
-            "conversations.list",
-            {"types": "public_channel"},
-        )
-        channel = next(
-            item for item in channels["channels"] if item["name"] == "reborn-alerts"
-        )
-        await slack_post(
-            client, emulate_url, "conversations.join", {"channel": channel["id"]}
-        )
-        await slack_post(
-            client,
-            emulate_url,
-            "chat.postMessage",
-            {"channel": channel["id"], "text": "QA10 self-authored earlier message"},
-        )
-        await slack_post(
-            client,
-            emulate_url,
-            "chat.postMessage",
-            {
-                "channel": channel["id"],
-                "text": "ENTITYMSG_1784643032040 QA10 searchable marker",
-            },
-        )
-        root = await slack_post(
-            client,
-            emulate_url,
-            "chat.postMessage",
-            {"channel": channel["id"], "text": "QA10 thread root"},
-        )
-        await slack_post(
-            client,
-            emulate_url,
-            "chat.postMessage",
-            {
-                "channel": channel["id"],
-                "thread_ts": root["ts"],
-                "text": "QA10 visible thread reply",
-            },
-        )
-    return {
-        "team_id": identity["team_id"],
-        "user_id": identity["user_id"],
-        "reviewer_id": by_name["qa-reviewer"]["id"],
-        "channel_id": channel["id"],
-        "channel_name": channel["name"],
-        "thread_ts": root["ts"],
-    }
-
-
-async def _configure_slack(base_url: str, slack_state: dict[str, str]) -> None:
-    client_id = "reborn-qa-emulate-slack-client"
-    async with httpx.AsyncClient(headers=reborn_bearer_headers()) as client:
-        configured = await client.get(
-            f"{base_url}/api/webchat/v2/operator/extension-configuration",
-            timeout=30,
-        )
-        configured.raise_for_status()
-        group = next(
-            item
-            for item in configured.json()["groups"]
-            if item["group_id"] == "extension.slack"
-        )
-        response = await client.put(
-            f"{base_url}/api/webchat/v2/operator/extension-configuration/extension.slack",
-            json={
-                "values": [
-                    {"handle": "slack_bot_token", "value": EMULATE_SLACK_BEARER},
-                    {"handle": "slack_signing_secret", "value": "emulate-signing-secret"},
-                    {"handle": "slack_team_id", "value": slack_state["team_id"]},
-                    {"handle": "slack_api_app_id", "value": client_id},
-                    {"handle": "slack_installation_id", "value": slack_state["team_id"]},
-                    {"handle": "slack_bot_user_id", "value": slack_state["user_id"]},
-                    {"handle": "slack_oauth_client_id", "value": client_id},
-                    {
-                        "handle": "slack_oauth_client_secret",
-                        "value": "emulate-slack-client-secret",
-                    },
-                ],
-                "expected_revision": group["revision"],
-                "idempotency_key": f"reborn-qa-emulate-{uuid.uuid4()}",
-            },
-            timeout=30,
-        )
-        response.raise_for_status()
-        assert response.json()["complete"] is True, response.text
-
-
-async def _seed_github_account(base_url: str) -> None:
-    async with httpx.AsyncClient(headers=reborn_bearer_headers()) as client:
-        response = await client.post(
-            f"{base_url}/api/webchat/v2/extensions/github/setup",
-            json={
-                "action": "submit",
-                "client_action_id": client_action_id(),
-                "payload": {
-                    "secrets": {"github_runtime_token": EMULATE_GITHUB_BEARER},
-                    "fields": {},
-                },
-            },
-            timeout=30,
-        )
-        response.raise_for_status()
-        secret = next(
-            item
-            for item in response.json()["secrets"]
-            if item["name"] == "github_runtime_token"
-        )
-        assert secret["provided"] is True, response.text
-
-
-async def _seed_slack_account(
-    base_url: str,
-    emulate_url: str,
-    slack_state: dict[str, str],
-) -> None:
-    expires_at = (datetime.now(UTC) + timedelta(minutes=5)).isoformat()
-    async with httpx.AsyncClient(headers=reborn_bearer_headers()) as client:
-        requirement = await fetch_extension_oauth_requirement(
-            client,
-            base_url,
-            "slack",
-        )
-        started = await client.post(
-            f"{base_url}/api/webchat/v2/extensions/slack/setup/oauth/start",
-            json={
-                "requirement": requirement["name"],
-                "expires_at": expires_at,
-                "invocation_id": requirement["setup"].get("invocation_id"),
-            },
-            timeout=30,
-        )
-        assert started.is_success, started.text
-        body = started.json()
-        query = parse_qs(urlparse(body["authorization_url"]).query)
-        consent = await client.post(
-            f"{emulate_url}/oauth/v2/authorize/callback",
-            data={
-                "user_id": slack_state["user_id"],
-                "redirect_uri": query["redirect_uri"][0],
-                "scope": query.get("scope", [""])[0],
-                "user_scope": query.get("user_scope", [""])[0],
-                "state": query["state"][0],
-                "client_id": query["client_id"][0],
-            },
-            follow_redirects=False,
-            timeout=30,
-        )
-        assert consent.status_code == 302, consent.text
-        callback_query = parse_qs(urlparse(consent.headers["location"]).query)
-        callback = await client.get(
-            f"{base_url}/api/reborn/product-auth/oauth/slack/callback",
-            params={key: values[0] for key, values in callback_query.items()},
-            headers={"Accept": "application/json"},
-            timeout=30,
-        )
-        assert callback.is_success, callback.text
-        flow_status = await client.get(
-            f"{base_url}/api/reborn/product-auth/oauth/flow/{body['flow_id']}/status",
-            params={"invocation_id": body["callback_scope"]["invocation_id"]},
-            timeout=30,
-        )
-        flow_status.raise_for_status()
-        assert flow_status.json()["status"] == "completed", flow_status.text
-
-
-async def _install_extensions(base_url: str, extension_ids: tuple[str, ...]) -> None:
-    async with httpx.AsyncClient(headers=reborn_bearer_headers()) as client:
-        for extension_id in extension_ids:
-            installed = await client.post(
-                f"{base_url}/api/webchat/v2/extensions/install",
-                json={
-                    "package_ref": {"kind": "extension", "id": extension_id},
-                    "client_action_id": client_action_id(),
-                },
-                timeout=30,
-            )
-            installed.raise_for_status()
-
-
-async def _assert_extensions_active(
-    base_url: str, extension_ids: tuple[str, ...]
-) -> None:
-    async with httpx.AsyncClient(headers=reborn_bearer_headers()) as client:
-        listed = await client.get(
-            f"{base_url}/api/webchat/v2/extensions",
-            timeout=30,
-        )
-        listed.raise_for_status()
-        by_id = {
-            extension["package_ref"]["id"]: extension
-            for extension in listed.json()["extensions"]
-        }
-        for extension_id in extension_ids:
-            extension = by_id.get(extension_id)
-            assert extension is not None, (
-                f"{extension_id} disappeared after completing its manifest-declared setup"
-            )
-            assert extension["installation_state"] == "active", extension
-
-
-def _provider_leg(trace: dict, provider_tools: frozenset[str]) -> dict:
-    """Keep the recorded provider decisions and final response in order."""
-    provider_steps = []
-    final_text = None
-    for step in trace["steps"][1:]:
-        response = step["response"]
-        if response["type"] == "tool_calls":
-            calls = [
-                call
-                for call in response["tool_calls"]
-                if call["name"] in provider_tools
-                or call["name"] in TRACE_BOOTSTRAP_TOOLS
-            ]
-            if calls:
-                provider_steps.append(
-                    {"response": {"type": "tool_calls", "tool_calls": calls}}
-                )
-        elif response["type"] == "text":
-            final_text = response
-
-    assert provider_steps, "provider journey must retain at least one tool call"
-    assert final_text is not None, "provider journey must retain a final response"
-    return {
-        **trace,
-        "steps": [trace["steps"][0], *provider_steps, {"response": final_text}],
-    }
-
-
-def _result_binding(tool_call_id: str, pointer: str) -> dict:
-    return {
-        "$trace_result": {
-            "tool_call_id": tool_call_id,
-            "pointer": pointer,
-        }
-    }
-
-
-def _inject_deferred_tool_disclosure(trace: dict) -> None:
-    """Translate the harvested extension flow to today's deferred-tool flow."""
-    provider_names = list(
-        dict.fromkeys(
-            call["name"]
-            for step in trace["steps"]
-            for call in step["response"].get("tool_calls", [])
-            if call["name"] in PROVIDER_TOOL_NAMES
-        )
+def _compile_journey_case(journey_case, slack_state: dict[str, str]):
+    trace_path = ROOT / journey_case.trace
+    return compile_provider_journey_trace(
+        load_recorded_trace(trace_path),
+        source=trace_path.name,
+        facts=journey_case.replay,
+        provider_tools=PROVIDER_TOOL_NAMES,
+        slack_state=slack_state,
     )
-    first_provider_step = next(
-        index
-        for index, step in enumerate(trace["steps"])
-        if any(
-            call["name"] in PROVIDER_TOOL_NAMES
-            for call in step["response"].get("tool_calls", [])
-        )
-    )
-    trace["steps"].insert(
-        first_provider_step,
-        {
-            "response": {
-                "type": "tool_calls",
-                "tool_calls": [
-                    {
-                        "id": f"call_disclose_{index}",
-                        "name": "capability_info",
-                        "arguments": {"name": name.replace("__", ".")},
-                    }
-                    for index, name in enumerate(provider_names)
-                ],
-            }
-        },
-    )
-
-
-def test_injected_deferred_tool_disclosures_have_stable_ids():
-    trace = {
-        "steps": [
-            {"response": {"type": "user_input", "content": "inspect Slack"}},
-            {
-                "response": {
-                    "type": "tool_calls",
-                    "tool_calls": [
-                        {
-                            "id": "call_whoami",
-                            "name": "slack__whoami",
-                            "arguments": {},
-                        }
-                    ],
-                }
-            },
-            {"response": {"type": "text", "content": "done"}},
-        ]
-    }
-
-    _inject_deferred_tool_disclosure(trace)
-
-    disclosure_calls = trace["steps"][1]["response"]["tool_calls"]
-    assert disclosure_calls == [
-        {
-            "id": "call_disclose_0",
-            "name": "capability_info",
-            "arguments": {"name": "slack.whoami"},
-        }
-    ]
-
-
-def _coalesce_independent_provider_reads(trace: dict, batch_size: int = 25) -> None:
-    """Fit QA 9C's independent Slack reads within today's loop-turn limit."""
-    provider_indexes = [
-        index
-        for index, step in enumerate(trace["steps"])
-        if any(
-            call["name"] in PROVIDER_TOOL_NAMES
-            for call in step["response"].get("tool_calls", [])
-        )
-    ]
-    calls = [
-        call
-        for index in provider_indexes
-        for call in trace["steps"][index]["response"]["tool_calls"]
-    ]
-    insertion_index = provider_indexes[0]
-    trace["steps"] = [
-        step
-        for index, step in enumerate(trace["steps"])
-        if index not in provider_indexes
-    ]
-    batches = [
-        {
-            "response": {
-                "type": "tool_calls",
-                "tool_calls": calls[start : start + batch_size],
-            }
-        }
-        for start in range(0, len(calls), batch_size)
-    ]
-    trace["steps"][insertion_index:insertion_index] = batches
-
-
-def _normalize_google_arguments(trace: dict, case: str) -> None:
-    created_document_call_id = None
-    created_spreadsheet_call_id = None
-    document_upload_contents = []
-    for step in trace["steps"]:
-        for call in step["response"].get("tool_calls", []):
-            if call["name"] == "google-docs__create_document":
-                document_upload_contents.append("")
-            elif call["name"] == "google-docs__insert_text":
-                for index, content in enumerate(document_upload_contents):
-                    if content == "":
-                        document_upload_contents[index] = call["arguments"].get("text", "")
-                        break
-    document_upload_index = 0
-    seeded_spreadsheet = (
-        "sheet_reborn_bug_tracker" if case.startswith("qa_7") else "sheet_reborn_abc"
-    )
-    seeded_sheet_id = 0
-
-    for step in trace["steps"]:
-        if "tool_calls" not in step["response"]:
-            continue
-        normalized_calls = []
-        for call in step["response"].get("tool_calls", []):
-            name = call["name"]
-            arguments = call["arguments"]
-            _replace_value(arguments, "EMAIL_REDACTED", "e2e.google@example.com")
-
-            if name == "google-docs__create_document":
-                created_document_call_id = call.get("id")
-                call["name"] = "google-drive__upload_file"
-                content = (
-                    document_upload_contents[document_upload_index]
-                    if document_upload_index < len(document_upload_contents)
-                    else ""
-                )
-                document_upload_index += 1
-                call["arguments"] = {
-                    "name": arguments["title"],
-                    "content": content,
-                    "mime_type": "text/plain",
-                }
-            elif name == "google-docs__insert_text":
-                continue
-            elif name.startswith("google-docs__") and "document_id" in arguments:
-                call["name"] = "google-drive__download_file"
-                call["arguments"] = {
-                    "file_id": (
-                        _result_binding(created_document_call_id, "/file/id")
-                        if created_document_call_id
-                        else "drv_near_ai_strategy"
-                    )
-                }
-
-            if name == "google-sheets__create_spreadsheet":
-                created_spreadsheet_call_id = call.get("id")
-            elif name.startswith("google-sheets__"):
-                if "spreadsheet_id" in arguments:
-                    arguments["spreadsheet_id"] = (
-                        _result_binding(
-                            created_spreadsheet_call_id, "/spreadsheet_id"
-                        )
-                        if created_spreadsheet_call_id
-                        else seeded_spreadsheet
-                    )
-                if "sheet_id" in arguments:
-                    arguments["sheet_id"] = (
-                        _result_binding(
-                            created_spreadsheet_call_id,
-                            "/sheets/0/sheet_id",
-                        )
-                        if created_spreadsheet_call_id
-                        else seeded_sheet_id
-                    )
-
-            if name == "gmail__get_message":
-                arguments["message_id"] = "msg_emulate_near_inbound"
-            elif name == "google-drive__download_file":
-                arguments["file_id"] = "drv_pepsico_account_brief"
-            normalized_calls.append(call)
-        step["response"]["tool_calls"] = normalized_calls
-    trace["steps"] = [
-        step
-        for step in trace["steps"]
-        if step["response"].get("type") != "tool_calls"
-        or step["response"].get("tool_calls")
-    ]
-
-
-def test_google_normalization_seeds_consistent_spreadsheet_and_sheet_ids():
-    trace = {
-        "steps": [
-            {
-                "response": {
-                    "type": "tool_calls",
-                    "tool_calls": [
-                        {
-                            "id": "call_format",
-                            "name": "google-sheets__format_cells",
-                            "arguments": {
-                                "spreadsheet_id": "harvested-spreadsheet",
-                                "sheet_id": 123,
-                            },
-                        }
-                    ],
-                }
-            }
-        ]
-    }
-
-    _normalize_google_arguments(trace, "qa_7c_slack_bug_logger_routine")
-
-    assert trace["steps"][0]["response"]["tool_calls"][0]["arguments"] == {
-        "spreadsheet_id": "sheet_reborn_bug_tracker",
-        "sheet_id": 0,
-    }
-
-
-def _normalize_slack_arguments(
-    trace: dict, slack_state: dict[str, str], case: str
-) -> None:
-    for step in trace["steps"]:
-        for call in step["response"].get("tool_calls", []):
-            if not call["name"].startswith("slack__"):
-                continue
-            arguments = call["arguments"]
-            if "channel" in arguments:
-                arguments["channel"] = (
-                    "C_REBORN_QA_10E_MISSING"
-                    if case == "qa_10e_slack_error_honesty"
-                    else slack_state["channel_id"]
-                )
-            if "user_id" in arguments:
-                arguments["user_id"] = slack_state["reviewer_id"]
-            if "thread_ts" in arguments:
-                arguments["thread_ts"] = slack_state["thread_ts"]
-            if "text" in arguments:
-                arguments["text"] = arguments["text"].replace(
-                    "SLACK_ID_REDACTED", slack_state["reviewer_id"]
-                )
-            if "query" in arguments:
-                arguments["query"] = arguments["query"].replace(
-                    "SLACK_ID_REDACTED", slack_state["channel_name"]
-                )
-
-
-def _replace_value(value: object, old: str, new: str) -> None:
-    if isinstance(value, dict):
-        for key, child in value.items():
-            if child == old:
-                value[key] = new
-            else:
-                _replace_value(child, old, new)
-    elif isinstance(value, list):
-        for index, child in enumerate(value):
-            if child == old:
-                value[index] = new
-            else:
-                _replace_value(child, old, new)
-
-
-async def _load_trace(
-    mock_llm_server: str,
-    trace_path: Path,
-    *,
-    provider_tools: frozenset[str] | None = None,
-    slack_state: dict[str, str] | None = None,
-) -> dict:
-    trace = json.loads(trace_path.read_text(encoding="utf-8"))
-    if provider_tools is not None:
-        trace = _provider_leg(trace, provider_tools)
-    if trace_path.stem == "qa_9c_slack_digest_names_not_ids":
-        _coalesce_independent_provider_reads(trace)
-        _inject_deferred_tool_disclosure(trace)
-    elif trace_path.stem == "qa_10e_slack_error_honesty":
-        trace["steps"][-1]["request_hint"] = {
-            "expected_failed_tool_result_contains": "channel_not_found"
-        }
-    if provider_tools is not None:
-        _normalize_google_arguments(trace, trace_path.stem)
-    if slack_state is not None:
-        _normalize_slack_arguments(trace, slack_state, trace_path.stem)
-    async with httpx.AsyncClient() as client:
-        response = await client.post(
-            f"{mock_llm_server}/__mock/llm_trace",
-            json={"source": trace_path.name, "trace": trace},
-            timeout=15,
-        )
-        response.raise_for_status()
-    return trace
 
 
 async def _install_inline_trace(
@@ -896,9 +238,7 @@ async def _install_inline_trace(
     response.raise_for_status()
 
 
-def _provider_operation_trace(
-    case: ProviderOperationCase, arguments: dict
-) -> dict:
+def _provider_operation_trace(case: ProviderOperationCase, arguments: dict) -> dict:
     wire_name = capability_id_to_wire_name(case.capability_id)
     return {
         "steps": [
@@ -1003,277 +343,6 @@ async def _fetch_all_timeline_pages_with_retry(
         seen_cursors.add(cursor)
 
 
-def _recorded_provider_calls(trace: dict) -> list[dict]:
-    return [
-        call
-        for step in trace["steps"]
-        for call in step["response"].get("tool_calls", [])
-        if call["name"] in PROVIDER_TOOL_NAMES
-    ]
-
-
-def _google_created_resource_call(calls: list[dict]) -> dict | None:
-    return next(
-        (
-            call
-            for call in calls
-            if call["name"]
-            in {
-                "google-docs__create_document",
-                "google-drive__upload_file",
-                "google-sheets__create_spreadsheet",
-            }
-        ),
-        None,
-    )
-
-
-def _google_created_resource_name(call: dict) -> str:
-    if call["name"] == "google-drive__upload_file":
-        return call["arguments"]["name"]
-    return call["arguments"]["title"]
-
-
-def _raw_trace_uses_tool_prefix(trace_path: Path, prefix: str) -> bool:
-    trace = json.loads(trace_path.read_text(encoding="utf-8"))
-    return any(
-        call["name"].startswith(prefix)
-        for step in trace["steps"]
-        for call in step["response"].get("tool_calls", [])
-    )
-
-
-async def _emulate_github_supports_release_writes(emulate_url: str) -> bool:
-    async with httpx.AsyncClient(headers=github_headers(), timeout=15) as client:
-        response = await client.post(
-            f"{emulate_url}/repos/nearai/ironclaw/releases",
-            json=GITHUB_RELEASE_WRITE_PROBE_PAYLOAD,
-        )
-    if response.status_code in GITHUB_RELEASE_WRITE_UNAVAILABLE:
-        return False
-    if response.status_code != 422:
-        raise AssertionError(
-            "GitHub release write probe must reject the invalid payload "
-            f"without mutation; got {response.status_code}: {response.text}"
-        )
-    return True
-
-
-async def _assert_google_provider_outcome(
-    emulate_url: str, case: str, trace: dict
-) -> None:
-    calls = _recorded_provider_calls(trace)
-    async with httpx.AsyncClient(headers=google_headers(), timeout=15) as client:
-        if case in {
-            "qa_2f_calendar_prep_email_delivery",
-            "qa_4e_github_release_email_delivery",
-        }:
-            send = next(call for call in calls if call["name"] == "gmail__send_message")
-            subject = send["arguments"]["message"]["subject"]
-            listed = await client.get(
-                f"{emulate_url}/gmail/v1/users/me/messages",
-                params={"q": f"subject:{subject}"},
-            )
-            listed.raise_for_status()
-            assert listed.json().get("messages"), f"sent message missing for {case}"
-
-        create_call = _google_created_resource_call(calls)
-        if create_call is None:
-            return
-
-        title = _google_created_resource_name(create_call)
-        files = await client.get(
-            f"{emulate_url}/drive/v3/files",
-            params={"q": f"name = '{title}' and trashed = false"},
-        )
-        files.raise_for_status()
-        matching = [item for item in files.json()["files"] if item["name"] == title]
-        assert matching, f"created Google resource missing for {case}: {files.text}"
-        resource_id = matching[-1]["id"]
-
-        if create_call["name"] == "google-drive__upload_file":
-            media = await client.get(
-                f"{emulate_url}/drive/v3/files/{resource_id}",
-                params={"alt": "media"},
-            )
-            media.raise_for_status()
-            assert media.text == create_call["arguments"].get("content", ""), media.text
-            return
-
-        if create_call["name"] == "google-docs__create_document":
-            document = await client.get(f"{emulate_url}/v1/documents/{resource_id}")
-            document.raise_for_status()
-            assert "QA5D-NONCE" in document.text, document.text
-            return
-
-        spreadsheet = await client.get(
-            f"{emulate_url}/v4/spreadsheets/{resource_id}"
-        )
-        spreadsheet.raise_for_status()
-        if case == "qa_6e_gmail_to_sheet_delivery":
-            values = await client.get(
-                f"{emulate_url}/v4/spreadsheets/{resource_id}/values/Sheet1!A1:C10"
-            )
-            values.raise_for_status()
-            assert "REBORN_QA_6E_GMAIL_TO_SHEET_DELIVERY" in values.text, values.text
-        elif case == "qa_7e_slack_bug_sheet_delivery":
-            values = await client.get(
-                f"{emulate_url}/v4/spreadsheets/{resource_id}/values/Bugs!A1:E10"
-            )
-            values.raise_for_status()
-            assert "REBORN_QA_7E_BUG_ROW" in values.text, values.text
-
-
-async def _assert_google_provider_baseline(
-    emulate_url: str, case: str, trace: dict
-) -> None:
-    """Prove this journey cannot observe mutations from an earlier journey."""
-    calls = _recorded_provider_calls(trace)
-    async with httpx.AsyncClient(headers=google_headers(), timeout=15) as client:
-        send = next(
-            (call for call in calls if call["name"] == "gmail__send_message"),
-            None,
-        )
-        if send is not None:
-            subject = send["arguments"]["message"]["subject"]
-            listed = await client.get(
-                f"{emulate_url}/gmail/v1/users/me/messages",
-                params={"q": f"subject:{subject}"},
-            )
-            listed.raise_for_status()
-            assert not listed.json().get("messages"), (
-                f"provider world for {case} already contains sent mail {subject!r}"
-            )
-
-        create_call = _google_created_resource_call(calls)
-        if create_call is None:
-            return
-        title = _google_created_resource_name(create_call)
-        files = await client.get(
-            f"{emulate_url}/drive/v3/files",
-            params={"q": f"name = '{title}' and trashed = false"},
-        )
-        files.raise_for_status()
-        assert not [item for item in files.json()["files"] if item["name"] == title], (
-            f"provider world for {case} already contains Google resource {title!r}"
-        )
-
-
-async def _assert_slack_provider_outcome(
-    emulate_url: str,
-    slack_state: dict[str, str],
-    trace: dict,
-) -> None:
-    sends = [
-        call
-        for call in _recorded_provider_calls(trace)
-        if call["name"] == "slack__send_message"
-    ]
-    if not sends:
-        return
-    async with httpx.AsyncClient(timeout=15) as client:
-        for send in sends:
-            messages = await _slack_messages_for_send(
-                client, emulate_url, slack_state, send
-            )
-            assert any(
-                message.get("text") == send["arguments"]["text"]
-                for message in messages
-            ), messages
-
-
-async def _slack_messages_for_send(
-    client: httpx.AsyncClient,
-    emulate_url: str,
-    slack_state: dict[str, str],
-    send: dict,
-) -> list[dict]:
-    """Read a delivery from the Slack surface that can contain it."""
-    payload = {"channel": slack_state["channel_id"], "limit": 100}
-    thread_ts = send["arguments"].get("thread_ts")
-    if thread_ts is None:
-        method = "conversations.history"
-    else:
-        method = "conversations.replies"
-        payload["ts"] = thread_ts
-    page = await slack_post(client, emulate_url, method, payload)
-    return page.get("messages", [])
-
-
-async def _assert_slack_provider_baseline(
-    emulate_url: str,
-    slack_state: dict[str, str],
-    case: str,
-    trace: dict,
-) -> None:
-    """Prove an earlier journey did not leave the expected Slack mutation."""
-    sends = [
-        call
-        for call in _recorded_provider_calls(trace)
-        if call["name"] == "slack__send_message"
-    ]
-    if not sends:
-        return
-    async with httpx.AsyncClient(timeout=15) as client:
-        for send in sends:
-            messages = await _slack_messages_for_send(
-                client, emulate_url, slack_state, send
-            )
-            assert not any(
-                message.get("text") == send["arguments"]["text"]
-                for message in messages
-            ), f"provider world for {case} already contains the expected Slack delivery"
-
-
-async def _cleanup_slack_provider_mutations_from_trace(
-    emulate_url: str,
-    slack_state: dict[str, str],
-    trace: dict,
-) -> None:
-    sends = [
-        call
-        for call in _recorded_provider_calls(trace)
-        if call["name"] == "slack__send_message"
-    ]
-    if not sends:
-        return
-
-    async with httpx.AsyncClient(timeout=15) as client:
-        matches = {}
-        for send in sends:
-            messages = await _slack_messages_for_send(
-                client, emulate_url, slack_state, send
-            )
-            expected_text = send["arguments"]["text"]
-            matches.update(
-                {
-                    message["ts"]: message
-                    for message in messages
-                    if message.get("text") == expected_text
-                }
-            )
-        for message in matches.values():
-            await slack_post(
-                client,
-                emulate_url,
-                "chat.delete",
-                {"channel": slack_state["channel_id"], "ts": message["ts"]},
-            )
-
-
-async def _cleanup_slack_provider_mutations(
-    emulate_url: str,
-    slack_state: dict[str, str],
-    case: str,
-) -> None:
-    """Remove messages created by one journey without rotating OAuth state."""
-    trace = json.loads((TRACE_DIR / f"{case}.json").read_text(encoding="utf-8"))
-    _normalize_slack_arguments(trace, slack_state, case)
-    await _cleanup_slack_provider_mutations_from_trace(
-        emulate_url, slack_state, trace
-    )
-
-
 async def test_slack_mutation_cleanup_covers_thread_replies(
     resettable_emulate_provider_world,
 ) -> None:
@@ -1326,18 +395,13 @@ async def test_slack_mutation_cleanup_covers_thread_replies(
     }
     slack_state = {"channel_id": channel["id"]}
     try:
-        await _assert_slack_provider_outcome(emulate_url, slack_state, trace)
+        calls = recorded_provider_calls(trace, PROVIDER_TOOL_NAMES)
+        await assert_slack_provider_outcome(emulate_url, slack_state, calls)
         with pytest.raises(AssertionError, match="already contains"):
-            await _assert_slack_provider_baseline(
-                emulate_url, slack_state, "thread-cleanup-contract", trace
-            )
+            await assert_slack_provider_baseline(emulate_url, slack_state, calls)
 
-        await _cleanup_slack_provider_mutations_from_trace(
-            emulate_url, slack_state, trace
-        )
-        await _assert_slack_provider_baseline(
-            emulate_url, slack_state, "thread-cleanup-contract", trace
-        )
+        await cleanup_slack_provider_mutations(emulate_url, slack_state, calls)
+        await assert_slack_provider_baseline(emulate_url, slack_state, calls)
     finally:
         async with httpx.AsyncClient(timeout=15) as client:
             for message in (reply, root):
@@ -1372,10 +436,10 @@ async def _replay_qa_journey_provider_leg(
     journey_case,
 ) -> None:
     """Replay one provider leg against the caller-selected provider lifecycle."""
-    case = journey_case.case_id
     server = provider_server["base_url"]
     trace_path = ROOT / journey_case.trace
-    if _raw_trace_uses_tool_prefix(trace_path, "google-sheets__"):
+    recorded_trace = load_recorded_trace(trace_path)
+    if recorded_trace_uses_tool_prefix(recorded_trace, "google-sheets__"):
         async with httpx.AsyncClient(headers=google_headers(), timeout=15) as client:
             emulate_google_url = provider_server["emulate_google_url"]
             response = await client.get(
@@ -1383,45 +447,43 @@ async def _replay_qa_journey_provider_leg(
             )
         if response.status_code == 404:
             pytest.skip("Emulate 0.7.0 does not expose the Google Sheets API")
-    if _raw_trace_uses_tool_prefix(trace_path, "slack__"):
+    if recorded_trace_uses_tool_prefix(recorded_trace, "slack__"):
         async with httpx.AsyncClient(headers=slack_headers(), timeout=15) as client:
             emulate_slack_url = provider_server["emulate_slack_url"]
             response = await client.get(f"{emulate_slack_url}/api/auth.test")
         if response.status_code == 404:
             pytest.skip("Emulate 0.7.0 does not expose Slack Web API GET routes")
-    trace = await _load_trace(
-        mock_llm_server,
-        trace_path,
+    compiled = compile_provider_journey_trace(
+        recorded_trace,
+        source=trace_path.name,
+        facts=journey_case.replay,
         provider_tools=PROVIDER_TOOL_NAMES,
         slack_state=provider_server["slack_state"],
     )
+    trace = compiled.trace
+    await _install_inline_trace(mock_llm_server, compiled.source, trace)
     user_input = trace["steps"][0]["response"]["content"]
-    expected_calls = _recorded_provider_calls(trace)
+    expected_calls = recorded_provider_calls(trace, PROVIDER_TOOL_NAMES)
 
-    await _assert_google_provider_baseline(
-        provider_server["emulate_google_url"], case, trace
+    await assert_google_provider_baseline(
+        provider_server["emulate_google_url"], expected_calls
     )
-    await _assert_slack_provider_baseline(
+    await assert_slack_provider_baseline(
         provider_server["emulate_slack_url"],
         provider_server["slack_state"],
-        case,
-        trace,
+        expected_calls,
     )
 
     async with httpx.AsyncClient(headers=reborn_bearer_headers()) as client:
         thread_id = await create_thread(client, server)
         await send_message(client, server, thread_id, user_input)
 
-        replay_timeout = 180 if case == "qa_9c_slack_digest_names_not_ids" else 120
-        replay = await _wait_for_trace_replay(
-            mock_llm_server, timeout=replay_timeout
-        )
+        replay_timeout = journey_case.replay.timeout_seconds
+        replay = await _wait_for_trace_replay(mock_llm_server, timeout=replay_timeout)
         assistant = await wait_for_assistant_message(
             client, server, thread_id, timeout=replay_timeout
         )
-        timeline = await _fetch_all_timeline_pages_with_retry(
-            client, server, thread_id
-        )
+        timeline = await _fetch_all_timeline_pages_with_retry(client, server, thread_id)
         previews = [
             preview
             for message in timeline.get("messages", [])
@@ -1436,34 +498,33 @@ async def _replay_qa_journey_provider_leg(
             if preview["capability_id"] in expected_counts
         )
         assert actual_counts == expected_counts, (actual_counts, expected_counts)
+        expected_failure = journey_case.replay.expected_capability_failure
         for preview in previews:
             if preview["capability_id"] not in expected_counts:
                 continue
             output = json.dumps(preview).lower()
-            if case == "qa_10e_slack_error_honesty":
+            if expected_failure is not None:
                 assert preview["status"] == "failed", json.dumps(preview)
-                assert "channel_not_found" in output, preview
+                assert expected_failure in output, preview
                 continue
             assert preview["status"] == "completed", json.dumps(preview)
             assert "auth_required" not in output, preview
             assert "not found" not in output, preview
 
-        if case == "qa_2c_drive_connect":
-            assert "Secondary Private Brief" not in json.dumps(timeline)
-        elif case == "qa_10e_slack_error_honesty":
-            assert "channel_not_found" in assistant["content"]
+        if expected_failure is not None:
+            assert expected_failure in assistant["content"]
 
-    await _assert_google_provider_outcome(
-        provider_server["emulate_google_url"], case, trace
+    await assert_google_provider_outcome(
+        provider_server["emulate_google_url"], expected_calls
     )
-    await _assert_slack_provider_outcome(
+    await assert_slack_provider_outcome(
         provider_server["emulate_slack_url"],
         provider_server["slack_state"],
-        trace,
+        expected_calls,
     )
 
     assert replay == {
-        "source": trace_path.name,
+        "source": compiled.source,
         "next_response": len(trace["steps"]) - 1,
         "response_count": len(trace["steps"]) - 1,
         "complete": True,
@@ -1506,10 +567,16 @@ async def test_mutating_qa_journeys_replay_in_reverse_against_shared_provider_wo
                         str(world) == "slack"
                         for world in journey_case.mutable_provider_worlds
                     ):
-                        await _cleanup_slack_provider_mutations(
+                        compiled = _compile_journey_case(
+                            journey_case,
+                            reborn_qa_emulate_runtime["slack_state"],
+                        )
+                        await cleanup_slack_provider_mutations(
                             reborn_qa_emulate_runtime["emulate_slack_url"],
                             reborn_qa_emulate_runtime["slack_state"],
-                            journey_case.case_id,
+                            recorded_provider_calls(
+                                compiled.trace, PROVIDER_TOOL_NAMES
+                            ),
                         )
         finally:
             if reset_services:
@@ -1533,7 +600,7 @@ async def test_provider_operation_case_executes_with_provider_readback(
     ]
     if (
         operation_case.provider_service == "github"
-        and not await _emulate_github_supports_release_writes(emulate_url)
+        and not await emulate_github_supports_release_writes(emulate_url)
     ):
         pytest.skip("Selected Emulate GitHub fixture does not expose repo write APIs")
     source = f"provider-operation-{operation_case.case_id}.json"
@@ -1552,9 +619,7 @@ async def test_provider_operation_case_executes_with_provider_readback(
         )
         replay = await _wait_for_trace_replay(mock_llm_server, timeout=120)
         await wait_for_assistant_message(client, server, thread_id, timeout=120)
-        timeline = await _fetch_all_timeline_pages_with_retry(
-            client, server, thread_id
-        )
+        timeline = await _fetch_all_timeline_pages_with_retry(client, server, thread_id)
 
     matches = [
         preview
@@ -1619,9 +684,7 @@ async def test_provider_fault_profile_preserves_safe_operation_outcomes(
         )
         replay = await _wait_for_trace_replay(mock_llm_server, timeout=120)
         await wait_for_assistant_message(client, server, thread_id, timeout=120)
-        timeline = await _fetch_all_timeline_pages_with_retry(
-            client, server, thread_id
-        )
+        timeline = await _fetch_all_timeline_pages_with_retry(client, server, thread_id)
 
     matches = [
         preview
@@ -1637,8 +700,7 @@ async def test_provider_fault_profile_preserves_safe_operation_outcomes(
     attempts = [
         attempt
         for attempt in proxy.state["requests"]
-        if attempt["method"] == fault_case.method
-        and attempt["path"] == fault_case.path
+        if attempt["method"] == fault_case.method and attempt["path"] == fault_case.path
     ]
     assert len(attempts) == 1, attempts
     assert attempts[0]["forwarded"] is fault_case.expected_forwarded
